+++ Runtime/Gameplay/Item/Equipment/EquipmentInstance.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+[Serializable]
+public class EquipmentInstance
+{
+    public EquipmentBase equipmentBase;
+    public List<StatModifier> statModifiers = new List<StatModifier>();
+
+    public EquipmentInstance(EquipmentBase equipment)
+    {
+        equipmentBase = equipment;
+        statModifiers = equipment.statModifiers.Select(m => m.CloneWithNewSource(this)).ToList();
+    }
+}