+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 0.0.3
+Addition of the equipment system
+
+Can create a scriptable object of an item (with stats), add it to an inventory, and equip an item.
+
+Example can be used by adding "Warrior" and "Canvas" from the example photo to the scene and drop a reference from warrior on the player reference on "Stat Debugger UI" on the canvas in scene.
+
+- Added: 
+    - Scriptable objects - ItemBase, EquipmentBase, Various equipment types (ArmorBase, JewelryBase, OffHandBase, WeaponBase)
+    - EquipmentManager, EquipmentInstance, EquipmentHelpers
+    - Some example equipment
+    - Basic UI for an inventory and equipment
+    - Stat UI
+    - Naughty Attributes (https://github.com/dbrizov/NaughtyAttributes)
+
 ## 0.0.2
 Project structure was changed, all folders came up one level.